+++ source/SpotInterruptionTriggerFunction/app.py
@@ -16,14 +16,18 @@
 import boto3
 import os
 import json
+import logging
 
 from botocore.exceptions import ClientError
+
+logger = logging.getLogger()
+logger.setLevel(logging.INFO)
 
 instance_metadata_table = boto3.resource('dynamodb').Table(os.environ['INSTANCE_METADATA_TABLE'])
 
 def lambda_handler(event, context):
 
-    print(event)
+    logger.info(event)
 
     # Transform CloudWatch Event
     item = {
@@ -31,12 +35,13 @@ def lambda_handler(event, context):
         'Region': event['region'],
         'LastEventTime': event['time'],
         'LastEventType': 'spot-interruption',
+        'State': 'none',
         'Interrupted': True,
         'InterruptedInstanceAction': event['detail']['instance-action'],
         'InterruptionTime': event['time']
     }
 
-    print(item)
+    logger.info(item)
 
     # Commit to DynamoDB
     try:
@@ -44,14 +49,15 @@ def lambda_handler(event, context):
             Key={
                 'InstanceId': item['InstanceId']
             },
-            UpdateExpression="SET #Region = :Region, #LastEventTime = :LastEventTime, #LastEventType = :LastEventType, #Interrupted = :Interrupted, #InterruptedInstanceAction = :InterruptedInstanceAction, #InterruptionTime = :InterruptionTime",
+            UpdateExpression="SET #Region = :Region, #LastEventTime = :LastEventTime, #LastEventType = :LastEventType, #Interrupted = :Interrupted, #InterruptedInstanceAction = :InterruptedInstanceAction, #InterruptionTime = :InterruptionTime, #EventHistory = list_append(if_not_exists(#EventHistory, :empty_list), :EventHistory)",
             ExpressionAttributeNames={
                 '#Region' : 'Region',
                 '#LastEventTime' : 'LastEventTime',
                 '#LastEventType' : 'LastEventType',
                 '#Interrupted' : 'Interrupted',
                 '#InterruptedInstanceAction' : 'InterruptedInstanceAction',
-                '#InterruptionTime' : 'InterruptionTime'
+                '#InterruptionTime' : 'InterruptionTime',
+                '#EventHistory' : 'EventHistory'
             },
             ExpressionAttributeValues={
                 ':Region': item['Region'],
@@ -59,15 +65,23 @@ def lambda_handler(event, context):
                 ':LastEventType': item['LastEventType'],
                 ':Interrupted': item['Interrupted'],
                 ':InterruptedInstanceAction': item['InterruptedInstanceAction'],
-                ':InterruptionTime': item['InterruptionTime']
+                ':InterruptionTime': item['InterruptionTime'],
+                ':EventHistory': [{ 
+                    "Name":  item['LastEventType'], 
+                    "Time": item['LastEventTime'],
+                    "State": item['State']
+                }],
+                ":empty_list": []
                 },
             ReturnValues="NONE"
         )
 
-        print(response)
+        logger.info(response)
     except ClientError as e:
-        print(e)
+        message = 'Error updating instance in DynamoDB: {}'.format(e)
+        logger.info(message)
+        raise Exception(message)
 
     # End
-    print('Execution Complete')
+    logger.info('Execution Complete')
     return