+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-11-18
+### Added
+- Support for capturing Rebalance Recommendation Events
+- Sample Athena Query for Rebalance Recommendation Events
+- Updated logging to use Logging library
+- Additional boto3 client exception handling
+- Add EventHistory attribute to DynamoDB so you can see and report on a history of all events
+- Adds aggregate interruption rate percentage to sample dashboard
+- Deployment button added to readme
+
 ## [1.0.0] - 2020-05-14
 ### Added
 - Initial Release